perf: improve the experience of redirection after login
