+++ console/src/router/guards/auth-check.ts
@@ -1,18 +1,18 @@
 import { useUserStore } from "@/stores/user";
-import type { Router } from "vue-router";
+import type { RouteLocationNormalized, Router } from "vue-router";
 
 const whiteList = ["Setup", "Login", "Binding"];
 
 export function setupAuthCheckGuard(router: Router) {
   router.beforeEach((to, from, next) => {
-    if (whiteList.includes(to.name as string)) {
-      next();
-      return;
-    }
-
     const userStore = useUserStore();
 
     if (userStore.isAnonymous) {
+      if (whiteList.includes(to.name as string)) {
+        next();
+        return;
+      }
+
       next({
         name: "Login",
         query: {
@@ -20,7 +20,48 @@ export function setupAuthCheckGuard(router: Router) {
         },
       });
       return;
+    } else {
+      if (to.name === "Login") {
+        if (allowRedirect(to)) {
+          next({
+            name: "Redirect",
+            query: {
+              redirect_uri: to.query.redirect_uri,
+            },
+          });
+        } else {
+          next({
+            name: "Dashboard",
+          });
+          return;
+        }
+      }
     }
+
     next();
   });
+}
+
+function allowRedirect(route: RouteLocationNormalized) {
+  const redirect_uri = route.query.redirect_uri as string;
+
+  if (!redirect_uri || redirect_uri === window.location.href) {
+    return false;
+  }
+
+  if (redirect_uri.startsWith("/")) {
+    return true;
+  }
+
+  if (
+    redirect_uri.startsWith("https://") ||
+    redirect_uri.startsWith("http://")
+  ) {
+    const url = new URL(redirect_uri);
+    if (url.origin === window.location.origin) {
+      return true;
+    }
+  }
+
+  return false;
 }
