+++ console/src/router/routes.config.ts
@@ -3,6 +3,7 @@ import NotFound from "@/views/exceptions/NotFound.vue";
 import Forbidden from "@/views/exceptions/Forbidden.vue";
 import BasicLayout from "@/layouts/BasicLayout.vue";
 import Setup from "@/views/system/Setup.vue";
+import Redirect from "@/views/system/Redirect.vue";
 import type { MenuGroupType } from "@halo-dev/console-shared";
 
 export const routes: Array<RouteRecordRaw> = [
@@ -29,6 +30,11 @@ export const routes: Array<RouteRecordRaw> = [
     meta: {
       title: "core.setup.title",
     },
+  },
+  {
+    path: "/redirect",
+    name: "Redirect",
+    component: Redirect,
   },
 ];
